+++ src/Patch/ServiceContainer/PatchExtension.php
@@ -25,6 +25,8 @@ use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
+use Meteor\Patch\Task\LimitBackupsHandler;
+use Meteor\Patch\Task\LimitBackups;
 
 class PatchExtension extends ExtensionBase implements ExtensionInterface, ScriptEventProviderInterface
 {
@@ -41,6 +43,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     const SERVICE_STRATEGY = 'patch.strategy';
     const SERVICE_TASK_BUS = 'patch.task_bus';
     const SERVICE_TASK_BACKUP_FILES_HANDLER = 'patch.task.backup_files_handler';
+    const SERVICE_TASK_LIMIT_BACKUPS_HANDLER = 'patch.task.limit_backups_handler';
     const SERVICE_TASK_CHECK_DATABASE_CONNECTION_HANDLER = 'patch.task.check_database_connection_handler';
     const SERVICE_TASK_CHECK_DISK_SPACE_HANDLER = 'patch.task.check_disk_space_handler';
     const SERVICE_TASK_CHECK_MODULE_CMS_DEPENDENCY_HANDLER = 'patch.task.check_module_cms_dependency_handler';
@@ -119,6 +122,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
 
         $this->loadBackupFinder($container);
         $this->loadBackupFilesTaskHandler($container);
+        $this->loadLimitBackupsTaskHandler($container);
         $this->loadCheckDatabaseConnectionTaskHandler($container);
         $this->loadCheckDiskSpaceHandler($container);
         $this->loadCheckModuleCmsDependencyTaskHandler($container);
@@ -141,7 +145,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadBackupFinder(ContainerBuilder $container)
     {
@@ -152,7 +156,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadTaskBus(ContainerBuilder $container)
     {
@@ -160,7 +164,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadBackupFilesTaskHandler(ContainerBuilder $container)
     {
@@ -177,7 +181,24 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
+     */
+    private function loadLimitBackupsTaskHandler(ContainerBuilder $container)
+    {
+        $definition = new Definition(LimitBackupsHandler::class, [
+            new Reference(self::SERVICE_BACKUP_FINDER),
+            new Reference(FilesystemExtension::SERVICE_FILESYSTEM),
+            new Reference(IOExtension::SERVICE_IO),
+        ]);
+        $definition->addTag(self::TAG_TASK_HANDLER, [
+            'task' => LimitBackups::class,
+        ]);
+
+        $container->setDefinition(self::SERVICE_TASK_LIMIT_BACKUPS_HANDLER, $definition);
+    }
+
+    /**
+     * @param ContainerBuilder $container
      */
     private function loadCheckDatabaseConnectionTaskHandler(ContainerBuilder $container)
     {
@@ -192,7 +213,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadCheckDiskSpaceHandler(ContainerBuilder $container)
     {
@@ -209,7 +230,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadCheckModuleCmsDependencyTaskHandler(ContainerBuilder $container)
     {
@@ -224,7 +245,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadCheckVersionTaskHandler(ContainerBuilder $container)
     {
@@ -240,7 +261,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadCheckWritePermissionTaskHandler(ContainerBuilder $container)
     {
@@ -255,7 +276,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadCopyFilesHandler(ContainerBuilder $container)
     {
@@ -272,7 +293,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadDeleteBackupHandler(ContainerBuilder $container)
     {
@@ -288,7 +309,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadDisplayVersionInfoTaskHandler(ContainerBuilder $container)
     {
@@ -304,7 +325,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadMigrateDownTaskHandler(ContainerBuilder $container)
     {
@@ -321,7 +342,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadMigrateUpTaskHandler(ContainerBuilder $container)
     {
@@ -337,7 +358,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadSetPermissionsHandler(ContainerBuilder $container)
     {
@@ -353,7 +374,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadUpdateMigrationVersionFilesTaskHandler(ContainerBuilder $container)
     {
@@ -370,7 +391,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadApplyCommand(ContainerBuilder $container)
     {
@@ -392,7 +413,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadVerifyCommand(ContainerBuilder $container)
     {
@@ -410,7 +431,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadManifestChecker(ContainerBuilder $container)
     {
@@ -420,7 +441,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadClearLockCommand(ContainerBuilder $container)
     {
@@ -438,7 +459,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadLocker(ContainerBuilder $container)
     {
@@ -446,7 +467,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadRollbackCommand(ContainerBuilder $container)
     {
@@ -469,7 +490,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadVersionInfoCommand(ContainerBuilder $container)
     {
@@ -485,7 +506,7 @@ class PatchExtension extends ExtensionBase implements ExtensionInterface, Script
     }
 
     /**
-     * @param ContailerBuilder $container
+     * @param ContainerBuilder $container
      */
     private function loadVersionComparer(ContainerBuilder $container)
     {