+++ tests/Migrations/Configuration/Fixtures/with_migrations/patch/upgrades/filesystem/Version0987654321.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Doctrine\Migrations;
+
+use Doctrine\DBAL\Migrations\AbstractMigration;
+use Doctrine\DBAL\Schema\Schema;
+
+class Version0987654321 extends AbstractMigration
+{
+    public function up(Schema $schema)
+    {
+    }
+
+    public function down(Schema $schema)
+    {
+    }
+}