+++ tests/Patch/Cli/Command/ApplyCommandTest.php
@@ -2,6 +2,7 @@
 
 namespace Meteor\Patch\Cli\Command;
 
+use InvalidArgumentException;
 use Meteor\Cli\Command\CommandTestCase;
 use Meteor\IO\NullIO;
 use Meteor\Patch\Event\PatchEvents;
@@ -25,6 +26,7 @@ class ApplyCommandTest extends CommandTestCase
         vfsStream::setup('root', null, [
             'patch' => [],
             'install' => [],
+            'logs' => []
         ]);
 
         $this->taskBus = Mockery::mock('Meteor\Patch\Task\TaskBusInterface');
@@ -125,7 +127,7 @@ class ApplyCommandTest extends CommandTestCase
     }
 
     /**
-     * @expectedException \InvalidArgumentException
+     * @expectedException InvalidArgumentException
      */
     public function testThrowsExceptionWhenWorkingDirIsTheSameAsTheInstallDir()
     {
@@ -485,6 +487,70 @@ class ApplyCommandTest extends CommandTestCase
             '--working-dir' => $workingDir,
             '--install-dir' => $installDir,
             '--skip-scripts' => null,
+        ]);
+    }
+
+    public function testChangesLogDirectoryIfPassedParameter()
+    {
+        $workingDir = vfsStream::url('root/patch');
+        $installDir = vfsStream::url('root/install');
+        $logDir = vfsStream::url('root/logs');
+        $filename = 'meteor-' . date('YmdHis') . '.log';
+
+        $config = ['name' => 'test'];
+        $this->command->setConfiguration($config);
+
+        $this->logger->shouldReceive('enable')
+            ->with($logDir.'/'.$filename)
+            ->once();
+
+        $this->manifestChecker->shouldReceive('check')
+            ->with($workingDir)
+            ->once();
+
+        $this->locker->shouldReceive('lock')
+            ->never();
+
+        $tasks = [new \stdClass()];
+        $this->strategy->shouldReceive('apply')
+            ->with($workingDir, $installDir, Mockery::any())
+            ->andReturn($tasks)
+            ->once();
+
+        $this->taskBus->shouldReceive('run')
+            ->with($tasks[0], $config)
+            ->andReturn(true)
+            ->once();
+
+        $this->locker->shouldReceive('unlock')
+            ->never();
+
+        $this->tester->execute([
+            '--working-dir' => $workingDir,
+            '--install-dir' => $installDir,
+            '--skip-lock' => null,
+            '--log-dir' => $logDir
+        ]);
+    }
+
+
+    public function testChangesLogDirectoryThrowsErrorIfDoesntExist()
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('The log directory `/testlog` does not exist.');
+
+        $workingDir = vfsStream::url('root/patch');
+        $installDir = vfsStream::url('root/install');
+
+        $config = ['name' => 'test'];
+        $this->command->setConfiguration($config);
+        
+
+        $this->tester->execute([
+            '--working-dir' => $workingDir,
+            '--install-dir' => $installDir,
+            '--skip-lock' => null,
+            '--log-dir' => '/testlog'
         ]);
     }
 }