+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 * Add support for 'replace_directories' configuration option, allowing certain folders to be marked for slightly different patch strategy whereby the existing folder is removed and replaced entirely by the folder from the package.
+* Updated Doctrine Migrations to 1.5.0 ([#110](https://github.com/jadu/meteor/pull/110)).
+* Fixed issues with running multiple scripts in the script runner
+* Adds --log-dir option at patching stage to specify a different folder for logs ([#84](https://github.com/jadu/meteor/pull/84)).
+* Adds a --limit-backups option at patching stage ([#83](https://github.com/jadu/meteor/pull/83)).
 * Adds correction to permission reset command in Troubleshooting documentation ([#94](https://github.com/jadu/meteor/pull/94)).
 
 ## v3.1.0